Redesign extreme-value family architecture

+++ src/Tail/MTail.jl
@@ -5,4 +5,4 @@ Corresponds to the MCopula viewed as an etreme value copula.
 """
 struct MTail <: Tail end
 A(::MTail, t::NTuple{d, <:Real}) where d = max(t)
-ExtremeValueCopula(d, ::MTail) = MCopula(d)
+ExtremeValueCopula(d::Int, ::MTail) = MCopula(d)
